+++ lib/draggable_panel.dart
@@ -1,3 +1,1 @@
 export 'src/draggable_panel.dart';
-export 'src/models/panel_button.dart';
-export 'src/models/panel_item.dart';