+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.4
+
+* The `DraggablePanelItem` and `DraggablePanelButton` models were removed and replaced with `Record`.
+This was done to make the package easier to use. If you were using `DraggablePanel` in a package, you would need to import models for others that use your same package. Now, this is not necessary.
+
 ## 0.0.3
 
 * Initial release.